Changed so that the app will give a toast when not started from launcher
Summary:
I just added a string to be printed to the toast
Then i moved the extras, that needs to be loadet to find stuff in the
launcher. to before a check that finds out whether or not the app is
connected to the launcher.
And i added the check.

Test Plan: None

Reviewers: aaaes13

Reviewed By: aaaes13

Subscribers: nula13

Differential Revision: http://web.giraf.cs.aau.dk/D55